requirements.txt
# Core dependencies
ui-tars>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Screen capture and control
pyautogui>=0.9.54
pillow>=10.1.0
mss>=9.0.1

# Async support
aiofiles>=23.2.1

# Configuration
pydantic>=2.0.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
